Added more error handling

+++ app/assets/js/pages/main.js
@@ -307,7 +307,7 @@ tools.bindEvent("click", ".startAbort .start-button:not([aria-disabled='true'])"
 tools.bindEvent("click", ".startAbort .abort-button:not([aria-disabled='true'])", () => {
     tools.setAborted(true);
 
-    if (tools.childProcess.pid) {
+    if (tools.childProcess && tools.childProcess.pid) {
         try {
             tools.childProcess.kill("SIGKILL");
             terminate(tools.childProcess.pid);
